Crate mailparsing

Structs

AddrSpec

AddressList

AttachmentOptions

AuthenticationResult

AuthenticationResults

Header

HeaderMap

Represents an ordered list of headers. Note that there may be multiple headers with the same name. Derefs to the underlying Vec<Header> for mutation, but provides some accessors for retrieving headers by name.

HeaderParseResult

Holds the result of parsing a block of headers

Mailbox

MailboxList

MessageBuilder

MessageConformance

MessageID

MimeParameters

MimePart

Parser

SimplifiedStructure

Enums

Address

ConformanceDisposition

ContentTransferEncoding

DecodedBody

MailParsingError

SharedString

Helper for holding either an owned or borrowed string, and where the slice method is aware of that borrowing, allowing for efficient copying and slicing without making extraneous additional copies

Traits

EncodeHeaderValue

Functions

has_lone_cr_or_lf

normalize_crlf

normalize_crlf_in_place

Type Aliases

Result