Crate mailparsing

Structs

ARCAuthenticationResults

AddrSpec

AddressList

AttachmentOptions

AuthenticationResult

AuthenticationResults

CheckFixSettings

Header

HeaderMap

Represents an ordered list of headers. Note that there may be multiple headers with the same name. Derefs to the underlying Vec<Header> for mutation, but provides some accessors for retrieving headers by name.

HeaderParseResult

Holds the result of parsing a block of headers

Mailbox

MailboxList

MessageBuilder

MessageConformance

MessageID

MimeParameters

MimePart

Parser

PartPointer

References the position of a MimePart by encoding the steps in a tree walking operation. The encoding of PartPointer is a sequence of integers that identify the index of a child part by its level within the mime tree, selecting the current node when no more indices remain. eg: [] indicates the root part, while [0] is the 0th child of the root.

Rfc2045Info

SimplifiedStructure

SimplifiedStructurePointers

Enums

Address

ConformanceDisposition

ContentTransferEncoding

DecodedBody

MailParsingError

SharedString

Helper for holding either an owned or borrowed string, and where the slice method is aware of that borrowing, allowing for efficient copying and slicing without making extraneous additional copies

Traits

EncodeHeaderValue

Functions

has_lone_cr_or_lf

normalize_crlf

normalize_crlf_in_place

Type Aliases

Result